+++ ItshfSim/mainwin.h
@@ -25,7 +25,7 @@ private slots:
     void on_new_time(const Time* ts);
     void on_new_state(int state, int dsec);
     void on_new_chan(int type, int glbChId, int snr, int n0, int regret);
-    void on_new_sts(int avgScan, int scanFrq, int linkNum, int testNum);
+    void on_new_sts(float avgScan, int scanFrq, int linkNum, int testNum);
 
     void on_scan_color(Qt::GlobalColor color);
     void on_link_color(Qt::GlobalColor color);
